nacl-binutils: Provide symbols for .note.gnu.build-id section bounds
This causes the linker to define (if referenced) symbols
__note_gnu_build_id_start and __note_gnu_build_id_end for the bounds
of the .note.gnu.build-id section.  This makes it possible for the
program to discover its own build ID, without having ELF program
headers accessible in the runtime memory image as they are with the
next-generation linkers.

BUG= https://code.google.com/p/nativeclient/issues/detail?id=3462
TEST= toolchain trybots
R=mseaborn@chromium.org

Review URL: https://codereview.chromium.org/16331012