Apply localmods other than mips backend to tree containing 2.24 branch
The 2.24 branch is in an entirely disconnected commit topology from
the 2.23 branch. So apply the sandboxing and bitcode linking-related
changes from 2.23 (the diff between our HEAD and the 7/21/13 daily
update commit from which we branched) to the corresponding commit
in the 2.24 commit tree.

This commit contains only non-autoconf-generated files.

R=jvoung@chromium.org
Review URL: https://codereview.chromium.org/253273005/