GAS: Handle multiple fixups in bundle for call adjustment
When using .bundle_lock, there can be multiple instructions, and thus
multiple fixups (relocs to generate) in a single frag.  When the frag
ends in a call instruction, all those fixups need to be adjusted for
the rearrangement of the instruction bytes done to move the call
instruction to the end of the bundle.

This also drops the ancient, unused, and unsupported handling of the
NACL_CONTROL_ENFORCE_RANGE environment variable.

BUG= https://code.google.com/p/nativeclient/issues/detail?id=4008
TEST= manual inspection of test case
R=dschuff@chromium.org, sehr@chromium.org

Review URL: https://codereview.chromium.org/773933002